Materials, point light and some updates to major stability of the engine

+++ DirectX_Game_Tut/Matrix4x4.h
@@ -97,10 +97,10 @@ public:
 			}
 			v.cross(vec[0], vec[1], vec[2]);
 
-			out.m_mat[0][i] = pow(-1.0f, i) * v.m_x / det;
-			out.m_mat[1][i] = pow(-1.0f, i) * v.m_y / det;
-			out.m_mat[2][i] = pow(-1.0f, i) * v.m_z / det;
-			out.m_mat[3][i] = pow(-1.0f, i) * v.m_w / det;
+			out.m_mat[0][i] = (float)pow(-1.0f, i) * v.m_x / det;
+			out.m_mat[1][i] = (float)pow(-1.0f, i) * v.m_y / det;
+			out.m_mat[2][i] = (float)pow(-1.0f, i) * v.m_z / det;
+			out.m_mat[3][i] = (float)pow(-1.0f, i) * v.m_w / det;
 		}
 
 		this->setMatrix(out);
